feat(scp): refresh console dashboard surfaces
## Overview
Move the SCP dashboard and ticket workspace onto the shared app shell while extracting the large dashboard page into focused components. The console now uses the refreshed design tokens, shared page-header slot, and renamed helper modules across dashboard, queue, search, preferences, and ticket detail pages.

## Problem
After introducing the shared panel shell, the SCP side still depended on the legacy DashboardLayout and kept most dashboard chart/rendering logic in one very large page component. Several SCP components also carried old token colors and lowercase helper filenames that no longer matched the repo's frontend naming direction.

Without this cleanup, the shell migration would leave two competing layout systems and make later dashboard or ticket-surface changes harder to isolate.

## Fix
- remove the legacy DashboardLayout and move remaining SCP pages to appShellLayout
- use PageHeader and SetPageHeader for dashboard, preferences, queue, search, and ticket-detail headers
- extract dashboard metrics, charts, header controls, section framing, recent activity, skeletons, and type helpers into resources/js/components/dashboard
- move LanguageSwitcher into the common component namespace for reuse by dashboard header controls
- rename sanitize-html to sanitizeHtml and update ticket body rendering imports
- rename queue-types to QueueTypes and centralize queue table, filter, pagination, and navigation types
- refresh SCP table, filter, avatar, ticket detail, and prose colors to the updated design token palette
- add expanded Tailwind theme tokens for primary, secondary, tertiary, neutral, success, warning, destructive foreground, and chart colors

## Verification
- git diff --check
- git diff --cached --check
- npx tsc --noEmit -p tsconfig.json
