--- resources/js/components/scp/queue-types.ts
@@ -1,16 +1,0 @@
-export interface QueueNode {
-    id: number;
-    title: string;
-    children?: QueueNode[];
-}
-
-export interface SavedSearch {
-    id: number;
-    title: string;
-}
-
-export interface QueueNavigation {
-    queues: QueueNode[];
-    personal: QueueNode[];
-    savedSearches: SavedSearch[];
-}
