+++ resources/js/pages/Scp/Preferences/Index.tsx
@@ -12,6 +12,7 @@ import {
 } from '@/components/ui/field';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
+import { PageHeader } from '@/components/layout/PageHeader';
 import {
     Select,
     SelectContent,
@@ -19,7 +20,7 @@ import {
     SelectTrigger,
     SelectValue,
 } from '@/components/ui/select';
-import DashboardLayout from '@/layouts/DashboardLayout';
+import { appShellLayout } from '@/layouts/AppShell';
 
 const THEME_OPTIONS = [
     { value: 'system', label: 'Match system' },
@@ -81,6 +82,7 @@ export default function PreferencesIndex({ preferences }: PreferencesPageProps) 
 
     return (
         <div className="mx-auto max-w-3xl">
+            <PageHeader title="Preferences" subtitle="Personalize how the agent workspace looks and notifies you." eyebrow="Settings" headerActions={null} />
             {props.status && (
                 <div className="mb-6 rounded-md border border-emerald-200 bg-emerald-50 px-4 py-3 text-sm text-emerald-900">
                     {props.status}
@@ -89,7 +91,7 @@ export default function PreferencesIndex({ preferences }: PreferencesPageProps) 
 
             <form
                 onSubmit={save}
-                className="rounded-[18px] border border-[#E2E8F0] bg-white p-6 shadow-sm shadow-[#0F172A]/[0.03] xl:p-8"
+                className="rounded-[18px] border border-[#E2E0D8] bg-white p-6 shadow-sm shadow-[#18181B]/[0.03] xl:p-8"
             >
                 <FieldGroup>
                     <Field>
@@ -142,18 +144,18 @@ export default function PreferencesIndex({ preferences }: PreferencesPageProps) 
                                 <Label
                                     key={id}
                                     htmlFor={`preferences-notify-${id}`}
-                                    className="flex items-start gap-3 rounded-md border border-[#E2E8F0] bg-[#F8FAFC] px-4 py-3 cursor-pointer hover:border-[#CBD5E1]"
+                                    className="flex items-start gap-3 rounded-md border border-[#E2E0D8] bg-[#FAFAF8] px-4 py-3 cursor-pointer hover:border-[#E2E0D8]"
                                 >
                                     <input
                                         id={`preferences-notify-${id}`}
                                         type="checkbox"
                                         checked={form.data.notifications[id]}
                                         onChange={(event) => toggleChannel(id, event.target.checked)}
-                                        className="mt-1 h-4 w-4 rounded border-[#CBD5E1] text-[#5B619D] focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-[#5B619D]"
+                                        className="mt-1 h-4 w-4 rounded border-[#E2E0D8] text-[#F97316] focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-[#F97316]"
                                     />
                                     <span className="flex flex-col gap-1">
-                                        <span className="text-sm font-medium text-[#0F172A]">{label}</span>
-                                        <span className="text-xs text-[#64748B]">{description}</span>
+                                        <span className="text-sm font-medium text-[#18181B]">{label}</span>
+                                        <span className="text-xs text-[#71717A]">{description}</span>
                                     </span>
                                 </Label>
                             ))}
@@ -175,14 +177,4 @@ type PreferencesPageComponent = typeof PreferencesIndex & {
     layout?: (page: ReactElement) => ReactNode;
 };
 
-(PreferencesIndex as PreferencesPageComponent).layout = (page: ReactElement) => (
-    <DashboardLayout
-        title="Preferences"
-        subtitle="Personalize how the agent workspace looks and notifies you."
-        eyebrow="Account"
-        activeNav="preferences"
-        headerActions={null}
-    >
-        {page}
-    </DashboardLayout>
-);
+(PreferencesIndex as PreferencesPageComponent).layout = appShellLayout;
