+++ resources/js/pages/Scp/Queues/Show.tsx
@@ -1,4 +1,4 @@
-import type { ReactNode } from 'react';
+import type { ReactElement, ReactNode } from 'react';
 import { HugeiconsIcon } from '@hugeicons/react';
 import {
     Add01Icon,
@@ -18,7 +18,7 @@ import {
     type QueueFilters,
     type QueueSortState,
 } from '@/components/scp/TicketFilterChips';
-import { type QueueNavigation } from '@/components/scp/queue-types';
+import { type QueueNavigation } from '@/components/scp/QueueTypes';
 import { Button, buttonVariants } from '@/components/ui/button';
 import {
     InputGroup,
@@ -26,7 +26,7 @@ import {
     InputGroupInput,
 } from '@/components/ui/input-group';
 import { Kbd } from '@/components/ui/kbd';
-import DashboardLayout from '@/layouts/DashboardLayout';
+import { appShellLayout, SetPageHeader } from '@/layouts/AppShell';
 
 interface Queue {
     id: number;
@@ -49,48 +49,8 @@ interface QueueShowLayoutProps extends QueueShowProps {
     children: ReactNode;
 }
 
-// Named function (not arrow) so Inertia treats it as a layout component, not a layout resolver.
-// Arrow functions have prototype===undefined, triggering a speculative call with raw props that
-// causes "TypeError: e.props is undefined". Named functions have a prototype, avoiding that path.
-function QueueShowLayout({ queue, navigation, children }: QueueShowLayoutProps) {
-    return (
-        <DashboardLayout
-            headerLeft={
-                <div>
-                    <QueueSelector navigation={navigation} activeQueueId={queue.id} variant="heading" label={queue.title} />
-                    <p className="mt-1 font-body text-sm text-[#94A3B8]">Tickets</p>
-                </div>
-            }
-            headerActions={
-                <>
-                    <a
-                        href={`/scp/queues/${queue.id}/export`}
-                        className={buttonVariants({ variant: 'outline', size: 'sm' })}
-                    >
-                        <HugeiconsIcon icon={Download01Icon} size={14} />
-                        Export CSV
-                    </a>
-                    <Button
-                        size="sm"
-                        disabled
-                        aria-disabled="true"
-                        title="Ticket creation lands in a later phase"
-                        className="rounded-md bg-emerald-500 text-white hover:bg-emerald-600 disabled:opacity-100"
-                    >
-                        <HugeiconsIcon icon={Add01Icon} size={14} />
-                        Add Ticket
-                    </Button>
-                </>
-            }
-            activeNav="queues"
-            contentClassName="w-full"
-        >
-            {children}
-        </DashboardLayout>
-    );
-}
-
 export default function QueueShow({
+    navigation,
     queue,
     tickets,
     pagination,
@@ -101,7 +61,30 @@ export default function QueueShow({
     sort,
 }: QueueShowProps) {
     return (
-        <div className="space-y-4">
+        <>
+            <SetPageHeader>
+                <div className="flex items-start justify-between gap-4">
+                    <div className="min-w-0 flex-1">
+                        <QueueSelector navigation={navigation} activeQueueId={queue.id} variant="heading" label={queue.title} />
+                        <p className="mt-1 font-body text-sm text-[#A1A1AA]">Tickets</p>
+                    </div>
+                    <div className="flex shrink-0 items-center gap-2 sm:gap-3">
+                        <a
+                            href={`/scp/queues/${queue.id}/export`}
+                            className={buttonVariants({ variant: 'outline', size: 'sm' })}
+                        >
+                            <HugeiconsIcon icon={Download01Icon} size={14} />
+                            Export CSV
+                        </a>
+                        <Button size="sm" disabled>
+                            <HugeiconsIcon icon={Add01Icon} size={14} />
+                            New Ticket
+                        </Button>
+                    </div>
+                </div>
+            </SetPageHeader>
+
+            <div className="space-y-4">
             {unsupported && (
                 <div className="rounded-[14px] border border-amber-200 bg-amber-50 px-5 py-4 text-sm text-amber-900">
                     <p className="font-medium">This queue uses legacy criteria that are not fully supported yet.</p>
@@ -115,7 +98,7 @@ export default function QueueShow({
                 </div>
             )}
 
-            <div className="flex flex-wrap items-center gap-x-6 gap-y-3 border-y border-zinc-100 px-1 py-3">
+            <div className="flex flex-wrap items-center gap-x-6 gap-y-3 border-y border-[#F4F2EB] px-1 py-3">
                 <form
                     role="search"
                     onSubmit={(event) => {
@@ -139,7 +122,7 @@ export default function QueueShow({
                     </InputGroup>
                 </form>
 
-                <span className="text-xs text-zinc-400">
+                <span className="text-xs text-[#A1A1AA]">
                     {pagination.total} {pagination.total === 1 ? 'ticket' : 'tickets'}
                 </span>
 
@@ -159,8 +142,9 @@ export default function QueueShow({
                 filters={filters}
             />
         </div>
+    </>
     );
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-(QueueShow as any).layout = QueueShowLayout;
+(QueueShow as any).layout = appShellLayout;
